Move Score dial to it's own page
