+++ frontend/src/components/HealthEngagementPage/HealthEngagementPage.styles.tsx
@@ -19,32 +19,6 @@ export const StyledHealthEngagementPage = styled.div`
         .highlight {
             color: ${(props) => props.theme.colors.primary};
         }
-        .circularProgressWithLabel {
-            width: 100%;
-            aspect-ratio: 1;
-            .MuiCircularProgress-root{
-                width: 100% !important;
-                height: auto !important;
-            }
-            
-            .progressForeground{
-                color: ${(props) => props.theme.colors.primary};
-                .MuiCircularProgress-circle{
-                    stroke-linecap:round;
-                }
-            }
-            .progressBackground{
-                color: ${(props) => props.theme.colors.surface};
-            }
-            .progressText{
-                color: ${(props) => props.theme.colors.text.secondary};
-                font-size: ${(props) => props.theme.typography.fontSize['5xl']};
-                font-weight: ${(props) => props.theme.typography.fontWeight.bold};
-    
-            }
-
-            
-        }
     }
     .fullWidthImage {
         width: 100vw;
